+++ lib/polyverse/openai/assistants.ts
@@ -5,18 +5,22 @@ import {
   Project,
   ProjectDataReference,
   Repository,
+  type PromptFileInfo,
 } from '../../data-model-types'
 import { getFileInfo } from '../backend/backend'
 import { isRecord } from '../typescript/helpers'
 import { submitTaskStepsAssistantFunction } from './assistantTools'
 import { OPENAI_MODEL } from './constants'
+import { mapPromptFileInfosToPromptFileTypes, type PromptFileTypes } from './../../../lib/polyverse/openai/utils'
+
 
 export const ASSISTANT_METADATA_CREATOR = 'sara.frontend'
 
 export const getVersion = () => {
-  // this should only be env variable or something runtime - not hardcoded
-  // TBD: To fix this, update deployment code to set the version globally
-  return '0.13.0'
+  // TODO: This should only be an env var or something set at runtime - not
+  // hardcoded. This needs to always be in sync with the versions listed in
+  // the `CHANGELOG.md` and `package.json` files
+  return '0.14.0'
 }
 
 interface FileTypes {
@@ -40,16 +44,37 @@ function createAssistantName(metadata: AssistantMetadata): string {
   return `${metadata.creator}-${metadata.version}-${metadata.orgName}-${metadata.userName}-${projectId}`
 }
 
-function getOpenAIAssistantInstructions(fileTypes: FileTypes): string {
+// Note that this function uses the new type `PromptFileTypes` which we have
+// designed as part of a data model/UI/Open AI logic refresh. Other functions
+// in this file use the old `FileTypes` type. Once we have fully cut over to
+// the new UI that consumes this new data model/Open AI logic  we ought to
+// update the other functions to use the new type.
+function getOpenAIAssistantInstructions(fileTypes: FileTypes | PromptFileTypes): string {
+  // This prompt was engineered to guide Sara on what she will be doing
+  // overall when she is created as an OpenAI Assistant. When specific questions
+  // are asked of her in a Thread and she is told to provide an answer to the
+  // question by performing a Thread Run each Thread Run ought to override the
+  // instructions if it will help Sara focus on the type of answer she ought to
+  // provide.
   return `
-    You are a coding assistant named Sara. 
-    You have access to the full codebase of a project in your files, including an ${fileTypes.aispec} file that summarizes the code. 
-    When asked a coding question, unless otherwise explicitly told not to, you give answers that use the relevant frameworks, APIs, data structures, and other aspects of the existing code. 
-    There are at least three files in your files that will help you answer questions. 
-    1. ${fileTypes.blueprint} is a very short summary of the overall architecture. It talks about what programming languages are used, major frameworks, and so forth. 
-    2. ${fileTypes.aispec} is another useful, medium size file. It has short summaries of all of the important code. 
-    3. ${fileTypes.projectsource} is the concatenation of all of the source code in the project. 
-    For all queries, use the ${fileTypes.blueprint} and ${fileTypes.aispec} files. Retrieve code snippets as needed from the concatenated code file ${fileTypes.projectsource}.`
+      You are a software architecture assistant as well as a coding assistant named Sara.
+      You have access to the full codebase of a project in your files, including a file named ${fileTypes.aispec} that summarizes the code.
+
+      Questions asked of you will be similiar to - but not exhaustive of - the bulleted list below:
+      * Questions focused on understanding and trying to define more granular work-items for high-level project goals
+      * Questions focused on trying to accomplish/complete a specific work-item that is associated with and required to complete a high-level project goal
+      * Questions focused on code/coding about the project where answers ought to use the relevant frameworks, APIs, data structures, and other aspects of the existing code
+      * Questions focused on software architecture and principals
+
+      If it is helpful you will be given additional details about how to answer specific types of questions when you go to answer them.
+
+      There are at least three files you have access to that will help you answer questions:
+      1. ${fileTypes.blueprint} is a very short summary of the overall architecture of the project. It talks about what programming languages are used, major frameworks, and so forth.
+      2. ${fileTypes.aispec} is another useful file that has short summaries of all of the important code in the project.
+      3. ${fileTypes.projectsource} is the concatenation of all of the source code in the project.
+
+      For all questions asked of you, use the ${fileTypes.blueprint} and ${fileTypes.aispec} files. Retrieve code snippets as needed from the concatenated code file ${fileTypes.projectsource}.
+  `
 }
 
 const oaiClient = new OpenAI({
@@ -195,4 +220,46 @@ export async function deleteAssistantFiles(assistant: Assistant) {
   )
 
   await Promise.all(fileDeletePromises)
+}
+
+// Note that this function uses the new type `PromptFileInfo` which we have
+// designed as part of a data model refresh. Other functions in this file use
+// the old `ProjectDataReference` type. Once we have fully cut over to the new
+// UI that consumes this new data model we ought to update the other functions
+// to use the new type.
+export const updateAssistantForPromptFileInfos = async (
+  promptFileInfos: PromptFileInfo[],
+  assistantMetadata: AssistantMetadata,
+): Promise<Assistant> => {
+  const assistant = await findAssistantFromMetadata(assistantMetadata)
+
+  if (!assistant) {
+    console.debug(
+      `Failed to find an assistant when updating one for prompt file infos using the following metadata: ${JSON.stringify(
+        assistantMetadata,
+      )}`,
+    )
+
+    throw new Error(
+      `Failed to find an assistant when updating one for prompt file infos`,
+    )
+  }
+
+  const fileIDs = promptFileInfos.map(({ id }) => id)
+  const identifiedPromptFileTypes =
+    mapPromptFileInfosToPromptFileTypes(promptFileInfos)
+
+  const prompt = getOpenAIAssistantInstructions(
+    identifiedPromptFileTypes,
+  )
+
+  return oaiClient.beta.assistants.update(assistant.id, {
+    file_ids: fileIDs,
+    instructions: prompt,
+    tools: [
+      { type: 'code_interpreter' },
+      { type: 'retrieval' },
+      submitTaskStepsAssistantFunction,
+    ],
+  })
 }